third_party/blink/web_tests/external/wpt/css/css-properties-values-api/animation/custom-property-animation-length-percentage-comma-list-expected.txt

This is a testharness.js-based test.
[FAIL] Animating a custom property of type <length-percentage># with iterationComposite
  assert_equals: expected "250%, calc(25% + 50px)" but got "250%, calc(25% + 250px)"
Harness: the test ran to completion.
